requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
requests
python-dotenv